Will/fix memory leak

+++ crates/brontes-pricing/src/graphs/all_pair_graph.rs
@@ -171,8 +171,8 @@ impl AllPairGraph {
                             let index = *self.token_to_index.get(&info.token_0).unwrap();
                             SubGraphEdge::new(
                                 PoolPairInfoDirection { info, token_0_in: node0 == index },
-                                i,
-                                path_length - i,
+                                i as u8,
+                                (path_length - i) as u8,
                             )
                         })
                         .collect_vec()
